agent-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
strands-agents==0.5.0
strands-agents-tools==0.3.0
boto3==1.34.0
pydantic==2.5.0
python-dotenv==1.0.0
tenacity==8.2.3
